chore: move changelog generation to publish
